% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{alinhar_por_posicao}
\alias{alinhar_por_posicao}
\title{Alinhar por posicao}
\usage{
alinhar_por_posicao(img)
}
\arguments{
\item{img}{imagem com uma coluna indicando a posicao do ponto.}
}
\description{
Alinhar por posicao
}

