% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{cortar}
\alias{cortar}
\title{Cortar}
\usage{
cortar(img, x_min = 10, x_max = 180, y_min = 12, y_max = 42)
}
\arguments{
\item{img}{imagem em formato de data.frame com as coluans x, y, r,g, e b}

\item{x_min}{10}

\item{x_max}{180}

\item{y_min}{12}

\item{y_max}{42}
}
\description{
Observando grande partes dos captchas vimos que a imagem está compreendida 
no intervalo: x >= 10, x <= 180 e y >= 12, y <= 42
}

