% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{decodificar}
\alias{decodificar}
\title{Decodificar}
\usage{
decodificar(arq, modelo, fun = sum)
}
\arguments{
\item{arq}{arquivo com o captcha}

\item{modelo}{modelo utilizado p/ prever}

\item{fun}{função usada p/ agregar os resultados.}
}
\description{
Decodificar
}

