% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{picotar}
\alias{picotar}
\title{Picotar}
\usage{
picotar(img, cortes = c(25, 55, 85, 120, 147))
}
\arguments{
\item{img}{imagem em formato de data.frame com as coluans x, y, r,g, e b}

\item{cortes}{vetor numérico de tamanho 5, com os pontos de corte}
}
\description{
Essa função serve p/ picotar a imagem nas 6 letras.
Ela recebe 5 pontos de corte p/ essa separação.
}

