% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{predizer}
\alias{predizer}
\title{Predizer}
\usage{
predizer(bd, modelo, fun = sum)
}
\arguments{
\item{bd}{banco de dados de uma imagem}

\item{modelo}{modelo utilizado p/ prever}

\item{fun}{função usada p/ agregar os resultados.}
}
\description{
Predizer
}

