% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{preparar}
\alias{preparar}
\title{Preparar banco de dados p/ modelo}
\usage{
preparar(dir, cortes = list(`1` = c(25, 55, 85, 120, 147), `2` = c(30, 55, 87,
  117, 145), `3` = c(27, 60, 87, 120, 148)))
}
\arguments{
\item{dir}{diretorio com todos os captchas já processados em data.frames}
}
\description{
Preparar banco de dados p/ modelo
}

