% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{processar}
\alias{processar}
\title{Processar Imagem}
\usage{
processar(img, cortes = c(25, 55, 85, 120, 147))
}
\arguments{
\item{img}{imagem em formato de data.frame com as colunas x, y, r,g, e b}

\item{cortes}{passado p/ a função \code{picotar}}
}
\description{
Processar Imagem
}

