% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{redimensionar}
\alias{redimensionar}
\alias{redimensionar_por_posicao}
\title{Redimensionar}
\usage{
redimensionar(img, w.out, h.out)

redimensionar_por_posicao(img, w.out = 20, h.out = 20)
}
\arguments{
\item{img}{imagem}

\item{w.out}{largura final}

\item{h.out}{altura final}
}
\description{
http://stackoverflow.com/questions/10865489/scaling-an-r-image

Redimensionar por grupo
}

