% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{separar}
\alias{separar}
\title{Separar em construção e validação}
\usage{
separar(bd, n_validacao = 100, seed = 500)
}
\arguments{
\item{bd}{banco de dados total.}

\item{n_validacao}{numero de casos na validação}

\item{seed}{semente usada}
}
\description{
Separar em construção e validação
}

